API

GlobalSearchRegression.export_csv — Method.

Exports main results with headers to file

source

GlobalSearchRegression.gsreg — Method.

gsreg(equation, data; kwargs...)

gsreg is a GlobalSearchRegression.jl function developed to perform all subset regressions on a set of potencial covariates, in order to select relevant features.

Basic Usage Example

julia> using GlobalSearchRegression, CSV, DataFrames
julia> data = CSV.read("path_to_your_data/your_data.csv")
julia> gsreg("y *", data)

Full-Syntax Example

julia> using GlobalSearchRegression
julia> gsreg("y x2 x3 x4 x5 x6 x7", data, 
    intercept=true, 
    outsample=10, 
    criteria=[:r2adj, :bic, :aic, :aicc, :cp, :rmse, :rmseout, :sse], 
    ttest=true, 
    method="precise", 
    vectoroperation=true,
    modelavg=true,
    residualtest=true,
    time=:x1,
    csv="output.csv",
    orderresults=false)

source

GlobalSearchRegression.get_data_position — Method.

Returns the position of the header value based on this structure. - Index - Covariates * b * bstd * T-test - Equation general information merged with criteria user-defined options. - Order from user combined criteria - Weight

source

GlobalSearchRegression.get_result_header — Method.

Constructs the header for results based in getdataposition orders.

source

GlobalSearchRegression.get_selected_cols — Method.

Returns selected appropiate covariates for each iteration

source