API

GlobalSearchRegression.export_csv — Method.

Exports main results with headers to file

GlobalSearchRegression.get_data_position — Method.

Returns the position of the header value based on this structure. - Index - Covariates * b * bstd * T-test - Equation general information merged with criteria user-defined options. - Order from user combined criteria - Weight

GlobalSearchRegression.get_result_header — Method.

Constructs the header for results based in getdataposition orders.

GlobalSearchRegression.get_selected_cols — Method.

Returns selected appropiate covariates for each iteration